+++ resources/views/admin/caseStudyData.blade.php
@@ -29,11 +29,14 @@
         <div class="block-header">
             <div class="row">
                 <div class="col-md-6 col-sm-12">
-                <h2>CASE STUDY DATA</h2>
+                    <h2>CASE STUDY DATA</h2>
                 </div>
-                <div class="col-md-6">
-                    <a href="{{ url('/add-case') }}" class="btn btn-primary float-right"><span><img src="{{ asset('assets/images/plus.png') }}" alt="All" class="add-icon"></span>Add</a>
-                </div>
+                @can('add-case-study')
+                    <div class="col-md-6">
+                        <a href="{{ url('/add-case') }}" class="btn btn-primary float-right"><span><img
+                                    src="{{ asset('assets/images/plus.png') }}" alt="All" class="add-icon"></span>Add</a>
+                    </div>
+                @endcan
             </div>
         </div>
         <div class="container-fluid">
@@ -55,7 +58,9 @@
                                             <th>Posted By</th>
                                             {{-- <th>Posted Date</th> --}}
                                             <th>Description</th>
-                                            <th>Action</th>
+                                            @if (auth()->user()->can('edit-case-study') || auth()->user()->can('delete-case-study'))
+                                                <th>Action</th>
+                                            @endif
                                         </tr>
 
                                     </thead>
@@ -68,7 +73,7 @@
                                                 <td><img src="{{ asset($data->avaDocs->path) }}"
                                                         style="width:70px;height:60px;border-radius:20%" /></td>
                                                 <td>{{ $data->posted_by }}</td>
-                                                <td>View 
+                                                <td>View
                                                     <i class="fa fa-eye" type="button" data-id="{{ $data->id }}"
                                                         data-toggle="modal" data-target="#exampleModalLong"
                                                         style="font-size:24px;cursor: pointer;"
@@ -76,15 +81,21 @@
                                                     </i>
 
                                                 </td>
-                                                <td>
-                                                    <div class="d-flex">
-                                                        <a href="{{ route('casestudy.edit', ['id' => $data->id]) }}"
-                                                            class="btn btn-primary">Edit</a>
-                                                        <button id="deleteButton"
-                                                            onclick="deleteModal('{{ $data->id }}')"
-                                                            class="btn btn-danger">Delete</button>
-                                                    </div>
-                                                </td>
+                                                @if (auth()->user()->can('edit-case-study') || auth()->user()->can('delete-case-study'))
+                                                    <td>
+                                                        <div class="d-flex">
+                                                            @can('edit-case-study')
+                                                                <a href="{{ route('casestudy.edit', ['id' => $data->id]) }}"
+                                                                    class="btn btn-primary">Edit</a>
+                                                            @endcan
+                                                            @can('delete-case-study')
+                                                                <button id="deleteButton"
+                                                                    onclick="deleteModal('{{ $data->id }}')"
+                                                                    class="btn btn-danger">Delete</button>
+                                                            @endcan
+                                                        </div>
+                                                    </td>
+                                                @endif
                                             </tr>
                                         @endforeach
                                     </tbody>
